// src/app/pages/model/CardDetails.ts
export class CardDetails {
  id: number;
  userId: number;
  companyLogo: string;
  companyName: string;
  contactNumber: string;
  whatsappNumber: string;
  email: string;
  website: string;
  aboutUsType: string;
  aboutUs: string;
  productsType: string;
  products: string;
  location: string;
  appointment: string;
  testimony: string;
  payMeName: string;
  payMeNumber: string;
  payMeCode: string;
  qrCode: string;
  facebook: string;
  instagram: string;
  linkedin: string;
  youtube: string;
  cardUserRole: string;
}