// src/app/pages/model/Roles.ts
export class Roles {
  id: number;
  userId: number;
  cardUserRole: string;
  userType: string;
  companyLogo: string;
  companyName: string;
  contactNumber: string;
  whatsappNumber: string;
  email: string;
  website: string;
  aboutUs: string;
  products: string;
  location: string;
  appointment: string;
  testimony: string;
  payMe: string;
  facebook: string;
  instagram: string;
  linkedin: string;
  youtube: string;
}